added wip code for level system
right now this code does not function correctly but I thought I would throw it in for the time being because it is a nice start for anyone to start a level system. This code will likely be replaced in the near future because I have some ideas to make it better.